requirements.txt
Flask
Flask-Session
Flask-Login
Flask-WTF
Flask-Bcrypt
Flask-Mail
Flask-Bootstrap5
flask-bootstrap5
Flask-Limiter
itsdangerous
WTForms
python-dotenv
PyMuPDF
python-docx
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
langchain-xai
gunicorn
flask_sqlalchemy
flask_migrate
flask_cors
sentence-transformers
openai
faiss-cpu

